# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Calculator Service Front End
    description: Service for multiplying numbers, a Goa teaser
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /multiply/{a}/{b}:
        get:
            tags:
                - calc
            summary: multiply calc
            operationId: calc#multiply
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
...